feat(xform): implement the hygienic macro transformer (pattern language)
Hygienic macros are implemented by adding a extended symbol type.
Local variables(not pattern variables) in templates are extended to
esymbol, bound a uid and the original env into it.

`syntax-rules` supports [srfi-149](https://srfi.schemers.org/srfi-149/srfi-149.html):
1. allow more than one consecutive <ellipsis> in template and
2. allow pattern variables in subtemplates to be followed by more instances
of the identifier <ellipsis> than they are followed in the subpattern in
which they occur.

And now core syntax keywords can be overwritten.